==> News_Articles/in_tibet_website/kbcmw/1_all_categories.ipynb <==
from bs4 import BeautifulSoup
import requests
import re


url = "http://ti.kbcmw.com"
headers = {
        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
    }
response = requests.get(url, headers=headers)
response.raise_for_status()

# Parse the page content with BeautifulSoup
soup = BeautifulSoup(response.content, 'html.parser')

categ_center = soup.find("ul", class_="cents")
# Find all 'a' tags
links = categ_center.find_all('a')
all_categories = []
# Extract and print the href attribute from each link
for link in links:
    href = link.get('href')
    if href:
        href = url + href
        text = link.get_text()
        all_categories.append((text, href))
        # print(href)

all_categories

def get_content(url):
    """
    
    """
    base_url = "http://ti.kbcmw.com"

    headers = {
        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
    }
    try:
        # Make the request to the URL
        response = requests.get(url, headers=headers)
        response.raise_for_status()
        
        # Parse the page content with BeautifulSoup
        soup = BeautifulSoup(response.content, 'html.parser')
        each_sub_cata = soup.find("div", id="xdh")
        all_new_link_text = []
        if each_sub_cata:
            all_links = each_sub_cata.find_all("a")
            if all_links:
                for each_link in all_links:
                    link = each_link.get('href')
                    href = base_url + link
                    text = each_link.get_text()
                    all_new_link_text.append([text, href])

                return all_new_link_text
            return None
        else:
            return None
    except Exception as e:
        print(e)



## extract sub categories

final_link = []
i = 0
new_categories = []
for each_link in all_categories[1:]:
    print(each_link)
    link = each_link[1]
    other_content = get_content(link)
    if other_content:
        print(other_content)
        new_categories.extend(other_content)
    else:
        new_categories.append(each_link)
        


len(new_categories)

## Extarct each categories page number

def get_page_no(url):
    """
    
    """

    headers = {
        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
    }
    try:
        # Make the request to the URL
        response = requests.get(url, headers=headers)
        response.raise_for_status()
        
        # Parse the page content with BeautifulSoup
        soup = BeautifulSoup(response.content, 'html.parser')
        page_block = soup.find("td", id="pagelist")
        if page_block:
            all_page_link = page_block.find_all("a", class_="num")
            if all_page_link:
                link = all_page_link[0].get('href')
                if link:
                    match = re.search(r'index_(\d+)\.html', link)
                    return [link, match.group(1)]
                    
        return ["no link last page link" , 0]
    except Exception as e:
        print(e)



url = "http://ti.kbcmw.com/html/news/china/"

get_page_no(url)

final_link = []
i = 0
for each_link in new_categories:
    link = each_link[1]
    page_no = get_page_no(link)
    clean_list = []
    clean_list.extend(each_link)
    clean_list.extend(page_no)
    # print(clean_list)
    final_link.append(clean_list)
    i = i+1
    if i//5 == 0:
        print(i)
        # break


final_link[:10]

import pandas as pd

df = pd.DataFrame(final_link, columns=['name_of_cate', 'original_url', 'Last_page_link', 'list_page_no'])
df.head()

df.to_csv("./data/All_cate.csv", index=False)

df.shape
