==> News_Articles/in_tibet_website/kbcmw/2_automate_all_categories.ipynb <==
import requests
from bs4 import BeautifulSoup
import json
import time

from tqdm import tqdm

import kbcmw_utils
import pandas as pd

## Here we will extract all the articles links from http://ti.kbcmw.com/ webpage
URL:  http://ti.kbcmw.com/

Tabs not extracted

 http://ti.kbcmw.com/


------------
-----------
------------
------------
------------

def loop_article_page(total_page, custom_url, key_code, base_url):
    """
    
    """
    return_file = {
        "Data": [],
        "message": "success",
        "response": 200
    }
    All_url_links = {}
    
    try:
        for i in tqdm(range(0, total_page+1)):
            if i == 0:
                final_url = custom_url + ".html"
            else:
                final_url = custom_url+ "_" + str(i) + ".html"
                
            found_url_links = kbcmw_utils.extract_kbcmw_page_article_links(final_url, base_url)
            key = key_code + str(i)
            All_url_links[key] = found_url_links
        return_file["Data"] = All_url_links
        return return_file
    
    except Exception as e:
        return_file["Data"] = All_url_links
        return_file["message"] = e
        return_file["response"] = 404
        return return_file

def check_error_in_links(All_url_link, page_code, print_each_error=False):
    """
    
    
    """

    error_counter = 0
    for page_id in range(1, len(All_url_link)):
        page_key = page_code + str(page_id)
        try:
            All_url_link.get(page_key)
            if  All_url_link.get(page_key)["Response"]!= 200:
                error_counter += 1
                if print_each_error:
                    print(page_key, All_url_link.get(page_key)["message"])
        except Exception as e:
            print(page_key, e)

    print(f"Total error in {page_code}: {error_counter}")

def save_json(path, file_name, data):
    """
    
    """
    with open(path+file_name, "w") as outfile:
        json.dump(data, outfile, indent=4, ensure_ascii=False)
        print(f"\n\nSuccessfully saved: {file_name}")

### all data

df = pd.read_csv("./data/All_cate.csv")
df.head(10)

sample_df = df
sample_df

for index, row in sample_df.iterrows():
    # print(row)
    name_of_cate = row["name_of_cate"]
    original_url = row["original_url"]
    Last_page_link = row["Last_page_link"]
    list_page_no = row["list_page_no"]

    total_page = list_page_no
    custom_url= original_url + "index"
    base_url = "http://ti.kbcmw.com"
    article_tag = name_of_cate
    key_code = "Page " + article_tag + " "
    all_links = loop_article_page(total_page, custom_url, key_code, base_url)
    
    print(f"Total page in {article_tag}: {len(all_links['Data'])}")
    check_error_in_links(all_links['Data'], key_code, print_each_error=True)

    # Saving the final file
    path = "./data/links/"
    file_name = f"tibetcm_ALL_link_{article_tag}.json"
    save_json(path, file_name, all_links['Data'])
