==> News_Articles/in_tibet_website/kbcmw/3_automate_all_content.ipynb <==
import requests
from bs4 import BeautifulSoup
import json
import time
from tqdm import tqdm
import re
import kbcmw_utils

import os
import json

import json
from bs4 import Tag
from tqdm import tqdm
import kbcmw_utils
import os

class CustomJSONEncoder(json.JSONEncoder):
    def default(self, obj):
        if isinstance(obj, Tag):
            return obj.get_text()  # Get all text from the tag
        return str(obj)  # Convert any other non-serializable object to string

def save_json(path, file_name, data):
    try: 
        with open(path + file_name, "w", encoding='utf-8') as outfile:
            json.dump(data, outfile, indent=4, ensure_ascii=False, cls=CustomJSONEncoder)
        print(f"Successfully saved: {file_name}")
    except Exception as e:
        print(f"Error saving {file_name}: {str(e)}")

def read_json(path, file_name):
    try:
        with open(path+file_name, 'r', encoding='utf-8') as openfile:
            Loaded_file = json.load(openfile)
            print(f"Successfully loaded: {file_name}")
        return Loaded_file
    except Exception as e:
        print(f"Error loading {file_name}: {str(e)}")
        return None

def get_content(All_links_data, Total_lenght, page_key_code, page_key_list):
    start = 0
    all_article = {}
    
    for i in tqdm(range(start, Total_lenght)):
        page_key = page_key_code + str(i)
        all_link_page = All_links_data[page_key]["Links"]
        page_article_index = 1
        
        for each_URL in all_link_page:
            try:
                article_content = kbcmw_utils.scrape_kbcmw_article_content(each_URL, tags=page_key_code)
                article_key = page_key + "_kbcmw_Article_" + str(page_article_index)
                page_article_index += 1
                all_article[article_key] = article_content
            except Exception as e:
                print(f"Error scraping {each_URL}: {str(e)}")
    
    Failure_count = 0
    for each_article_key in all_article:
        if all_article[each_article_key]["Response"] != 200:
            print(each_article_key)
            Failure_count += 1
    print(f"total Failure in the {page_key_list[1]} article: {Failure_count}")
    
    save_file_name = f"kbcmw_ALL_content_{page_key_list[1]}.json"
    print(save_file_name)
    path = "./data/all_content/"
    save_json(path, save_file_name, all_article)

def get_json_files(directory):
    json_files = []
    i = 0
    for filename in os.listdir(directory):
        if filename.endswith('.json'):
            file_path = os.path.join(directory, filename)
            try:
                with open(file_path, 'r', encoding='utf-8') as file:
                    i += 1
                    All_links_data = json.load(file)
                    Total_lenght = len(All_links_data)
                    print(f"Total page in {filename}: {Total_lenght}")
                
                    print(f"page key name: {list(All_links_data.keys())[-1]}")
                    page_key_list = list(All_links_data.keys())[-1].split(" ")
                    
                    page_key_code = "Page "+page_key_list[1]+" "
                    print(f"Page key code: {page_key_code}")
                    get_content(All_links_data, Total_lenght, page_key_code, page_key_list)
            except json.JSONDecodeError as e:
                print(f"Error decoding JSON in file {filename}: {str(e)}")
            except Exception as e:
                print(f"Error processing file {filename}: {str(e)}")
        # break
    print(f"Processed {i} files")


%%time
# Example usage
directory_path = './data/links/'
get_json_files(directory_path)
